+++ libs/media/src/lib/+state/media.service.ts
@@ -11,7 +11,8 @@ import { UploadData, HostedMediaFormValue, isValidMetadata } from "./media.fires
 
 // Blockframes
 import { UploadWidgetComponent } from "../file/upload-widget/upload-widget.component";
-import { delay, BehaviorStore } from "@blockframes/utils/helpers";
+import { delay } from "@blockframes/utils/helpers";
+import { BehaviorStore } from '@blockframes/utils/behavior-store';
 import { ImageParameters, getImgSize, getImgIxResourceUrl } from '../image/directives/imgix-helpers';
 import { clamp } from '@blockframes/utils/utils';
 import { deconstructFilePath, tempUploadDir, privacies, Privacy } from "@blockframes/utils/file-sanitizer";
